Add haversine distance formula
* Extract out all formulae to a separate mod
* Add haversine formula
* Add getters for latitude and longitude fields
* Change examples accordingly
